Make files properly respect the .gitignore
So if you try to .gitignore a file that's committed to the repository,
it won't work. This commit deletes the files from git's file system with
the command "git rm --cached [file]" and shows it here as deleting it
from the repo. It still exists on the local filesystem, however - it is
now treated exactly as a .gitignore'd file should be.